update seo elements
